+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 Changelog
 =========
 
+Version 0.2.1.0
+---------------
+
+*Jan 20, 2018*
+
+<https://github.com/mstksg/type-combinators-singletons/releases/tag/v0.2.1.0>
+
+*   Added `singSome`/`someSing` and `singWit1`/`wit1Sing`, to convert between
+    the *singletons* types `SomeSing` and `SingInstance` and their more generic
+    representations in *type-combinators*, `Some Sing` and `Wit1 Sing`.
+
 Version 0.2.0.0
 ---------------